ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
scikit-learn>=1.5.0
numpy>=2.0.0
pydantic>=2.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pytest>=7.0.0
